delete adjustment spectros as argument (#50)
* delete adjustment spectros as argument

* (temp) keep adjust_metadata.csv

---------

+++ src/OSmOSE/Spectrogram.py
@@ -7,6 +7,7 @@ from typing import List, Tuple, Union, Literal
 from math import log10
 from pathlib import Path
 import multiprocessing as mp
+from filelock import FileLock
 
 import pandas as pd
 import numpy as np
@@ -994,7 +995,7 @@ class Spectrogram(Dataset):
     # region On cluster
 
     def process_file(
-        self, audio_file: Union[str, Path], *, adjust: bool = False, save_matrix: bool = False
+        self, audio_file: Union[str, Path], *, adjust: bool = False, save_matrix: bool = False, clean_adjust_folder: bool = False
     ) -> None:
         """Read an audio file and generate the associated spectrogram.
 
@@ -1004,22 +1005,24 @@ class Spectrogram(Dataset):
             The name of the audio file to be processed
         adjust : `bool`, optional, keyword-only
             Indicates whether the file should be processed alone to adjust the spectrogram parameters (the default is False)
-        save_matrix : `save_matrix`, optional, keyword-only
+        save_matrix : `bool`, optional, keyword-only
             Whether to save the spectrogram matrices or not. Note that activating this parameter might increase greatly the volume of the project. (the default is False)
+        clean_adjust_folder: `bool`, optional, keyword-only
+            Whether the adjustment folder should be deleted.
         """
         set_umask()
 
         try:
-            if (self.path_output_spectrogram.parent.parent.joinpath(
+            if clean_adjust_folder and (self.path_output_spectrogram.parent.parent.joinpath(
                     "adjustment_spectros"
                 ).exists()
             ):
                 shutil.rmtree(
                     self.path_output_spectrogram.parent.parent.joinpath(
                         "adjustment_spectros"
-                    )
+                    ), ignore_errors=True
                 )
-        finally: 
+        except: 
             pass
 
         self.__build_path(adjust)
@@ -1303,17 +1306,20 @@ class Spectrogram(Dataset):
         metadata_input = self.path.joinpath(
             OSMOSE_PATH.spectrogram, "adjust_metadata.csv"
         )
+
         metadata_output = self.path.joinpath(
             OSMOSE_PATH.spectrogram,
             f"{str(self.spectro_duration)}_{str(self.sr_analysis)}",
             f"{str(self.nfft)}_{str(self.window_size)}_{str(self.overlap)}",
             "metadata.csv",
         )
-        try:
-            if not self.adjust and metadata_input.exists() and not metadata_output.exists():
-                metadata_input.rename(metadata_output)
-        finally:
-            pass
+        if not metadata_output.exists:
+            shutil.copyfile(metadata_input, metadata_output)
+        # try:
+        #     if not self.adjust and metadata_input.exists() and not metadata_output.exists():
+        #         metadata_input.rename(metadata_output)
+        # except:
+        #     pass
 
     # endregion
